[feat] complete Stage 2
Modified the program to read in a matrix of specified rows and columns, and multiply it by a specified constant.

Separated original matrix addition code in Main.java into its own method, then wrote a new method for the multiplication.

Added multiply method in Matrix.java for multiplication by a constant.